Move `quantum_theta` to `quantum.py` Rewrite tests for `quantum_theta`
