+++ plasmapy/formulary/tests/test_dimensionless.py
@@ -11,7 +11,6 @@ from plasmapy.formulary.dimensionless import (
     Hall_parameter,
     Mag_Reynolds,
     nD_,
-    quantum_theta,
     Re_,
     Reynolds_number,
     Rm_,
@@ -46,11 +45,6 @@ def test_aliases(alias, parent):
 def test_beta_dimensionless():
     # Check that beta is dimensionless
     float(beta(T, n, B))
-
-
-def test_quantum_theta_dimensionless():
-    # Check that quantum theta is dimensionless
-    float(quantum_theta(T, n))
 
 
 def test_beta_nan():
